+++ pkg/gio/v2/goutputstream.go
@@ -1167,7 +1167,7 @@ func (stream *OutputStream) Writev(ctx context.Context, vectors []OutputVector) 
 		_arg4 = (*C.GCancellable)(unsafe.Pointer(cancellable.Native()))
 	}
 	_arg2 = (C.gsize)(len(vectors))
-	_arg1 = (*C.GOutputVector)(C.malloc(C.size_t(len(vectors)) * C.size_t(C.sizeof_GOutputVector)))
+	_arg1 = (*C.GOutputVector)(C.malloc(C.size_t(uint(len(vectors)) * uint(C.sizeof_GOutputVector))))
 	defer C.free(unsafe.Pointer(_arg1))
 	{
 		out := unsafe.Slice((*C.GOutputVector)(_arg1), len(vectors))
@@ -1234,7 +1234,7 @@ func (stream *OutputStream) WritevAll(ctx context.Context, vectors []OutputVecto
 		_arg4 = (*C.GCancellable)(unsafe.Pointer(cancellable.Native()))
 	}
 	_arg2 = (C.gsize)(len(vectors))
-	_arg1 = (*C.GOutputVector)(C.malloc(C.size_t(len(vectors)) * C.size_t(C.sizeof_GOutputVector)))
+	_arg1 = (*C.GOutputVector)(C.malloc(C.size_t(uint(len(vectors)) * uint(C.sizeof_GOutputVector))))
 	defer C.free(unsafe.Pointer(_arg1))
 	{
 		out := unsafe.Slice((*C.GOutputVector)(_arg1), len(vectors))
@@ -1299,7 +1299,7 @@ func (stream *OutputStream) WritevAllAsync(ctx context.Context, vectors []Output
 		_arg4 = (*C.GCancellable)(unsafe.Pointer(cancellable.Native()))
 	}
 	_arg2 = (C.gsize)(len(vectors))
-	_arg1 = (*C.GOutputVector)(C.malloc(C.size_t(len(vectors)) * C.size_t(C.sizeof_GOutputVector)))
+	_arg1 = (*C.GOutputVector)(C.malloc(C.size_t(uint(len(vectors)) * uint(C.sizeof_GOutputVector))))
 	defer C.free(unsafe.Pointer(_arg1))
 	{
 		out := unsafe.Slice((*C.GOutputVector)(_arg1), len(vectors))
@@ -1413,7 +1413,7 @@ func (stream *OutputStream) WritevAsync(ctx context.Context, vectors []OutputVec
 		_arg4 = (*C.GCancellable)(unsafe.Pointer(cancellable.Native()))
 	}
 	_arg2 = (C.gsize)(len(vectors))
-	_arg1 = (*C.GOutputVector)(C.malloc(C.size_t(len(vectors)) * C.size_t(C.sizeof_GOutputVector)))
+	_arg1 = (*C.GOutputVector)(C.malloc(C.size_t(uint(len(vectors)) * uint(C.sizeof_GOutputVector))))
 	defer C.free(unsafe.Pointer(_arg1))
 	{
 		out := unsafe.Slice((*C.GOutputVector)(_arg1), len(vectors))
