+++ pkg/gio/v2/gtlsconnection.go
@@ -549,7 +549,7 @@ func (conn *TLSConnection) SetAdvertisedProtocols(protocols []string) {
 
 	_arg0 = (*C.GTlsConnection)(unsafe.Pointer(conn.Native()))
 	{
-		_arg1 = (**C.gchar)(C.malloc(C.size_t(len(protocols)+1) * C.size_t(unsafe.Sizeof(uint(0)))))
+		_arg1 = (**C.gchar)(C.malloc(C.size_t(uint((len(protocols) + 1)) * uint(unsafe.Sizeof(uint(0))))))
 		defer C.free(unsafe.Pointer(_arg1))
 		{
 			out := unsafe.Slice(_arg1, len(protocols)+1)
