+++ pkg/gtk/v4/gtkfilechooser.go
@@ -29,7 +29,7 @@ func init() {
 
 // FileChooserAction describes whether a GtkFileChooser is being used to open
 // existing files or to save to a possibly new file.
-type FileChooserAction int
+type FileChooserAction C.gint
 
 const (
 	// FileChooserActionOpen indicates open mode. The file chooser will only let
@@ -63,7 +63,7 @@ func (f FileChooserAction) String() string {
 
 // FileChooserError: these identify the various errors that can occur while
 // calling GtkFileChooser functions.
-type FileChooserError int
+type FileChooserError C.gint
 
 const (
 	// FileChooserErrorNonexistent indicates that a file does not exist.
@@ -266,7 +266,7 @@ func (chooser *FileChooser) AddChoice(id, label string, options, optionLabels []
 	_arg2 = (*C.char)(unsafe.Pointer(C.CString(label)))
 	defer C.free(unsafe.Pointer(_arg2))
 	{
-		_arg3 = (**C.char)(C.malloc(C.size_t(len(options)+1) * C.size_t(unsafe.Sizeof(uint(0)))))
+		_arg3 = (**C.char)(C.malloc(C.size_t(uint((len(options) + 1)) * uint(unsafe.Sizeof(uint(0))))))
 		defer C.free(unsafe.Pointer(_arg3))
 		{
 			out := unsafe.Slice(_arg3, len(options)+1)
@@ -279,7 +279,7 @@ func (chooser *FileChooser) AddChoice(id, label string, options, optionLabels []
 		}
 	}
 	{
-		_arg4 = (**C.char)(C.malloc(C.size_t(len(optionLabels)+1) * C.size_t(unsafe.Sizeof(uint(0)))))
+		_arg4 = (**C.char)(C.malloc(C.size_t(uint((len(optionLabels) + 1)) * uint(unsafe.Sizeof(uint(0))))))
 		defer C.free(unsafe.Pointer(_arg4))
 		{
 			out := unsafe.Slice(_arg4, len(optionLabels)+1)
